requirements.txt
ollama
requests
google-generativeai
